Added: Initial work on Text Alignments

+++ src/MysticChronicles.Engine/Objects/Element/Static/StaticText.cs
@@ -5,8 +5,8 @@ namespace MysticChronicles.Engine.Objects.Element.Static
 {
     public class StaticText : BaseElement
     {
-        private readonly int _xPosition;
-        private readonly int _yPosition;
+        private readonly float _xPosition;
+        private readonly float _yPosition;
         private readonly int _scale;
 
         private readonly string _text;
@@ -15,7 +15,7 @@ namespace MysticChronicles.Engine.Objects.Element.Static
 
         private readonly Color _color;
 
-        public StaticText(SpriteFont font, string text, Color color, int scale, int xPosition, int yPosition, bool isVisible = true)
+        public StaticText(SpriteFont font, string text, Color color, int scale, float xPosition, float yPosition, bool isVisible = true)
         {
             _font = font;
             _text = text;
